+++ packages/core/echo/echo-db/src/proxy-db/database.test.ts
@@ -7,6 +7,7 @@ import { afterEach, beforeEach, describe, expect, test } from 'vitest';
 
 import { Trigger } from '@dxos/async';
 import {
+  type BaseObject,
   create,
   dangerouslySetProxyId,
   Expando,
@@ -15,7 +16,7 @@ import {
   getType,
   type ReactiveObject,
 } from '@dxos/echo-schema';
-import { updateCounter } from '@dxos/echo-schema/testing';
+import { Contact, Container, RecordType, Task, updateCounter } from '@dxos/echo-schema/testing';
 import { registerSignalsRuntime } from '@dxos/echo-signals';
 import { PublicKey } from '@dxos/keys';
 import { openAndClose } from '@dxos/test-utils';
@@ -23,7 +24,7 @@ import { range } from '@dxos/util';
 
 import { getObjectCore, clone } from '../echo-handler';
 import { Filter } from '../query';
-import { Contact, Container, EchoTestBuilder, RecordType, Task } from '../testing';
+import { EchoTestBuilder } from '../testing';
 
 // TODO(burdon): Normalize tests to use common graph data (see query.test.ts).
 
@@ -425,7 +426,7 @@ describe('Database', () => {
     return { db, graph };
   };
 
-  const addToDatabase = async <T>(obj: ReactiveObject<T>) => {
+  const addToDatabase = async <T extends BaseObject>(obj: ReactiveObject<T>) => {
     const { db } = await createDbWithTypes();
     db.add(obj);
     await db.flush();