+++ packages/core/echo/echo-schema/src/mutable/mutable-schema.ts
@@ -12,11 +12,20 @@ import {
   updateFieldsInSchema,
   setTypenameInSchema,
 } from './manipulation';
-import { StoredSchema } from './types';
-import { type HasId, type JsonSchemaType, schemaVariance, type SchemaMeta, SchemaMetaSymbol } from '../ast';
+import { createStoredSchema, StoredSchema } from './types';
+import {
+  getObjectAnnotation,
+  schemaVariance,
+  type HasId,
+  type JsonSchemaType,
+  type SchemaMeta,
+  SchemaMetaSymbol,
+  type ObjectAnnotation,
+  EchoObject,
+  ObjectAnnotationId,
+} from '../ast';
 import { toEffectSchema, toJsonSchema } from '../json';
 
-// TODO(burdon): Reconcile with AbstractSchema.
 interface MutableSchemaConstructor extends S.Schema<MutableSchema> {
   new (): HasId;
 }
@@ -158,13 +167,19 @@ export class MutableSchema extends AbstractMutableSchema() implements S.Schema<a
     this._storedSchema.jsonSchema = toJsonSchema(removed);
   }
 
-  private _getSchema() {
+  /**
+   * Rebuilds this schema if it is dirty.
+   */
+  public rebuild() {
     if (this._isDirty || this._schema == null) {
       this._schema = toEffectSchema(unwrapProxy(this._storedSchema.jsonSchema));
       this._isDirty = false;
     }
+  }
 
-    return this._schema;
+  private _getSchema() {
+    this.rebuild();
+    return this._schema!;
   }
 }
 
@@ -193,4 +208,22 @@ const unwrapProxy = (jsonSchema: any): any => {
   }
 
   return result;
+};
+
+/**
+ * Create runtime representation of a schema.
+ */
+export const createMutableSchema = (
+  { typename, version }: ObjectAnnotation,
+  fields: S.Struct.Fields,
+): MutableSchema => {
+  const schema = S.partial(S.Struct(fields).omit('id')).pipe(EchoObject(typename, version));
+  const objectAnnotation = getObjectAnnotation(schema);
+  const schemaObject = createStoredSchema({ typename, version });
+  const updatedSchema = schema.annotations({
+    [ObjectAnnotationId]: { ...objectAnnotation, schemaId: schemaObject.id },
+  });
+
+  schemaObject.jsonSchema = toJsonSchema(updatedSchema);
+  return new MutableSchema(schemaObject);
 };