+++ packages/sdk/schema/src/testing/data.ts
@@ -16,6 +16,10 @@ import {
 
 import { createView, type ViewType } from '../view';
 
+/**
+ * @deprecated
+ */
+// TODO(burdon): Use Contact/Org, etc.
 export class TestSchema extends TypedObject({
   typename: 'example.com/type/Test',
   version: '0.1.0',