+++ packages/ui/lit-grid/src/testing/playwright/dx-grid.spec.ts
@@ -7,8 +7,8 @@ import { expect, test, type Page } from '@playwright/test';
 import { setupPage, storybookUrl } from '@dxos/test-utils/playwright';
 
 import { DxGridManager } from './dx-grid-manager';
-import { type DxGridCellsSelect } from '../types';
-import { toPlaneCellIndex } from '../util';
+import { type DxGridCellsSelect } from '../../types';
+import { toPlaneCellIndex } from '../../util';
 
 const gridPlaneCellSize = 31;
 const gap = 1;